+++ apps/api/src/graphql/accounts/account.resolvers.spec.ts
@@ -1,9 +1,9 @@
 import { AccountEntity } from '../../orm/entities/account.entity'
-import { AccountService } from './account.service'
 import { AccountResolvers } from './account.resolvers'
 import { Test } from '@nestjs/testing'
 import { AccountDto } from './account.dto'
 import { EthService } from '../../shared/eth.service'
+import { AccountService } from '../../dao/account.service'
 
 const address1 = '0000000000000000000000000000000000000001'
 const address2 = '0000000000000000000000000000000000000002'
